Deprecate the GTMHTTPFetcher.
If building with current tools and the min version is set high enough to use
NSURLSession, kicking errors/logs to migrate folks to the GTMSessionFetcher.

Apple is really only fixing things in NSURLSession at this point, and in
some of Google's apps, we've seen new issues show up in the older NSURConnection
code as it is no longer Apple's preferred networking layer.